+++ ruststep/tests/any.rs
@@ -54,7 +54,7 @@ fn deserialize_sub1() {
     test(
         "SUB_1(#3, 2.0)",
         Sub1Holder {
-            base: RValue::Entity(3).into(),
+            base: Name::Entity(3).into(),
             y1: 2.0,
         },
     );
@@ -82,7 +82,7 @@ fn deserialize_base_any() {
     test(
         "SUB_1(#3, 2.0)",
         BaseAnyHolder::Sub1(Box::new(Sub1Holder {
-            base: RValue::Entity(3).into(),
+            base: Name::Entity(3).into(),
             y1: 2.0,
         })),
     );
@@ -110,7 +110,7 @@ fn deserialize_base_any_placeholder() {
     test(
         "SUB_1(#3, 2.0)",
         PlaceHolder::Owned(BaseAnyHolder::Sub1(Box::new(Sub1Holder {
-            base: RValue::Entity(3).into(),
+            base: Name::Entity(3).into(),
             y1: 2.0,
         }))),
     );
@@ -162,18 +162,18 @@ fn into_base_any() {
 #[test]
 fn lookup_base_any() {
     test(
-        Parameter::RValue(RValue::Entity(1)),
+        Parameter::Ref(Name::Entity(1)),
         BaseAny::Base(Box::new(Base { x: 1.0 })),
     );
     test(
-        Parameter::RValue(RValue::Entity(2)),
+        Parameter::Ref(Name::Entity(2)),
         BaseAny::Sub1(Box::new(Sub1 {
             base: Base { x: 1.0 },
             y1: 2.0,
         })),
     );
     test(
-        Parameter::RValue(RValue::Entity(3)),
+        Parameter::Ref(Name::Entity(3)),
         BaseAny::Sub2(Box::new(Sub2 {
             base: Base { x: 1.0 },
             y2: 4.0,